@@ -1,6 +1,6 @@
 {
   "name": "detect-collisions",
-  "version": "3.0.10",
+  "version": "3.0.11",
   "description": "2d collision detection for circles, polygons, and points (with SAT and BVH)",
   "main": "dist/index.js",
   "types": "out/index.d.ts",
@@ -62,8 +62,8 @@
     "prettier": "^2.4.1",
     "ts-jest": "^27.0.7",
     "ts-loader": "^9.2.6",
-    "typescript": "^4.4.4",
-    "webpack": "^5.64.0",
+    "typescript": "^4.5.2",
+    "webpack": "^5.64.2",
     "webpack-cli": "^4.9.1"
   },
   "editions": [
