@@ -27,7 +27,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}))
 /***/ ((__unused_webpack_module, exports, __webpack_require__) => {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}));\nexports.Circle = undefined;\n\nvar _createClass = (function () {\n  function defineProperties(target, props) {\n    for (var i = 0; i < props.length; i++) {\n      var descriptor = props[i];\n      descriptor.enumerable = descriptor.enumerable || false;\n      descriptor.configurable = true;\n      if (\"value\" in descriptor) descriptor.writable = true;\n      Object.defineProperty(target, descriptor.key, descriptor);\n    }\n  }\n  return function (Constructor, protoProps, staticProps) {\n    if (protoProps) defineProperties(Constructor.prototype, protoProps);\n    if (staticProps) defineProperties(Constructor, staticProps);\n    return Constructor;\n  };\n})();\n\nvar _sat = __webpack_require__(/*! sat */ \"./node_modules/sat/SAT.js\");\n\nvar _sat2 = _interopRequireDefault(_sat);\n\nvar _model = __webpack_require__(/*! ../model */ \"./dist/model.js\");\n\nvar _utils = __webpack_require__(/*! ../utils */ \"./dist/utils.js\");\n\nfunction _interopRequireDefault(obj) {\n  return obj && obj.__esModule ? obj : { default: obj };\n}\n\nfunction _classCallCheck(instance, Constructor) {\n  if (!(instance instanceof Constructor)) {\n    throw new TypeError(\"Cannot call a class as a function\");\n  }\n}\n\nfunction _possibleConstructorReturn(self, call) {\n  if (!self) {\n    throw new ReferenceError(\n      \"this hasn't been initialised - super() hasn't been called\"\n    );\n  }\n  return call && (typeof call === \"object\" || typeof call === \"function\")\n    ? call\n    : self;\n}\n\nfunction _inherits(subClass, superClass) {\n  if (typeof superClass !== \"function\" && superClass !== null) {\n    throw new TypeError(\n      \"Super expression must either be null or a function, not \" +\n        typeof superClass\n    );\n  }\n  subClass.prototype = Object.create(superClass && superClass.prototype, {\n    constructor: {\n      value: subClass,\n      enumerable: false,\n      writable: true,\n      configurable: true,\n    },\n  });\n  if (superClass)\n    Object.setPrototypeOf\n      ? Object.setPrototypeOf(subClass, superClass)\n      : (subClass.__proto__ = superClass);\n}\n\n/**\n * collider - circle\n */\nvar Circle = (exports.Circle = (function (_SAT$Circle) {\n  _inherits(Circle, _SAT$Circle);\n\n  /**\n   * collider - circle\n   * @param {Vector} position {x, y}\n   * @param {number} radius\n   */\n  function Circle(position, radius) {\n    _classCallCheck(this, Circle);\n\n    var _this = _possibleConstructorReturn(\n      this,\n      (Circle.__proto__ || Object.getPrototypeOf(Circle)).call(\n        this,\n        (0, _utils.ensureVectorPoint)(position),\n        radius\n      )\n    );\n\n    _this.type = _model.Types.Circle;\n    _this.updateAABB();\n    return _this;\n  }\n  /**\n   * Updates Bounding Box of collider\n   */\n\n  _createClass(Circle, [\n    {\n      key: \"updateAABB\",\n      value: function updateAABB() {\n        this.minX = this.pos.x - this.r;\n        this.minY = this.pos.y - this.r;\n        this.maxX = this.pos.x + this.r;\n        this.maxY = this.pos.y + this.r;\n      },\n      /**\n       * Draws collider on a CanvasRenderingContext2D's current path\n       * @param {CanvasRenderingContext2D} context The canvas context to draw on\n       */\n    },\n    {\n      key: \"draw\",\n      value: function draw(context) {\n        var x = this.pos.x;\n        var y = this.pos.y;\n        var radius = this.r;\n        context.moveTo(x + radius, y);\n        context.arc(x, y, radius, 0, Math.PI * 2);\n      },\n    },\n  ]);\n\n  return Circle;\n})(_sat2.default.Circle));\n\n\n//# sourceURL=webpack://detect-collisions/./dist/bodies/circle.js?");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}));\nexports.Circle = undefined;\n\nvar _createClass = (function () {\n  function defineProperties(target, props) {\n    for (var i = 0; i < props.length; i++) {\n      var descriptor = props[i];\n      descriptor.enumerable = descriptor.enumerable || false;\n      descriptor.configurable = true;\n      if (\"value\" in descriptor) descriptor.writable = true;\n      Object.defineProperty(target, descriptor.key, descriptor);\n    }\n  }\n  return function (Constructor, protoProps, staticProps) {\n    if (protoProps) defineProperties(Constructor.prototype, protoProps);\n    if (staticProps) defineProperties(Constructor, staticProps);\n    return Constructor;\n  };\n})();\n\nvar _sat = __webpack_require__(/*! sat */ \"./node_modules/sat/SAT.js\");\n\nvar _sat2 = _interopRequireDefault(_sat);\n\nvar _model = __webpack_require__(/*! ../model */ \"./dist/model.js\");\n\nvar _utils = __webpack_require__(/*! ../utils */ \"./dist/utils.js\");\n\nfunction _interopRequireDefault(obj) {\n  return obj && obj.__esModule ? obj : { default: obj };\n}\n\nfunction _classCallCheck(instance, Constructor) {\n  if (!(instance instanceof Constructor)) {\n    throw new TypeError(\"Cannot call a class as a function\");\n  }\n}\n\nfunction _possibleConstructorReturn(self, call) {\n  if (!self) {\n    throw new ReferenceError(\n      \"this hasn't been initialised - super() hasn't been called\"\n    );\n  }\n  return call && (typeof call === \"object\" || typeof call === \"function\")\n    ? call\n    : self;\n}\n\nfunction _inherits(subClass, superClass) {\n  if (typeof superClass !== \"function\" && superClass !== null) {\n    throw new TypeError(\n      \"Super expression must either be null or a function, not \" +\n        typeof superClass\n    );\n  }\n  subClass.prototype = Object.create(superClass && superClass.prototype, {\n    constructor: {\n      value: subClass,\n      enumerable: false,\n      writable: true,\n      configurable: true,\n    },\n  });\n  if (superClass)\n    Object.setPrototypeOf\n      ? Object.setPrototypeOf(subClass, superClass)\n      : (subClass.__proto__ = superClass);\n}\n\n/**\n * collider - circle\n */\nvar Circle = (exports.Circle = (function (_SAT$Circle) {\n  _inherits(Circle, _SAT$Circle);\n\n  /**\n   * collider - circle\n   * @param {Vector} position {x, y}\n   * @param {number} radius\n   */\n  function Circle(position, radius) {\n    _classCallCheck(this, Circle);\n\n    var _this = _possibleConstructorReturn(\n      this,\n      (Circle.__proto__ || Object.getPrototypeOf(Circle)).call(\n        this,\n        (0, _utils.ensureVectorPoint)(position),\n        radius\n      )\n    );\n\n    _this.type = _model.Types.Circle;\n    _this.updateAABB();\n    return _this;\n  }\n  /**\n   * Updates Bounding Box of collider\n   */\n\n  _createClass(Circle, [\n    {\n      key: \"updateAABB\",\n      value: function updateAABB() {\n        this.minX = this.pos.x - this.r;\n        this.minY = this.pos.y - this.r;\n        this.maxX = this.pos.x + this.r;\n        this.maxY = this.pos.y + this.r;\n      },\n      /**\n       * Draws collider on a CanvasRenderingContext2D's current path\n       * @param {CanvasRenderingContext2D} context The canvas context to draw on\n       */\n    },\n    {\n      key: \"draw\",\n      value: function draw(context) {\n        var radius = this.r;\n        if (this.isTrigger) {\n          var max = radius / 2;\n          for (var i = 0; i < max; i++) {\n            var arc = (i / max) * 2 * Math.PI;\n            var arcPrev = ((i - 1) / max) * 2 * Math.PI;\n            var fromX = this.pos.x + Math.cos(arcPrev) * radius;\n            var fromY = this.pos.y + Math.sin(arcPrev) * radius;\n            var toX = this.pos.x + Math.cos(arc) * radius;\n            var toY = this.pos.y + Math.sin(arc) * radius;\n            (0, _utils.dashLineTo)(context, fromX, fromY, toX, toY);\n          }\n        } else {\n          context.moveTo(this.pos.x + radius, this.pos.y);\n          context.arc(this.pos.x, this.pos.y, radius, 0, Math.PI * 2);\n        }\n      },\n    },\n  ]);\n\n  return Circle;\n})(_sat2.default.Circle));\n\n\n//# sourceURL=webpack://detect-collisions/./dist/bodies/circle.js?");
 
 /***/ }),
 
@@ -49,7 +49,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}))
 /***/ ((__unused_webpack_module, exports, __webpack_require__) => {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}));\nexports.Polygon = undefined;\n\nvar _createClass = (function () {\n  function defineProperties(target, props) {\n    for (var i = 0; i < props.length; i++) {\n      var descriptor = props[i];\n      descriptor.enumerable = descriptor.enumerable || false;\n      descriptor.configurable = true;\n      if (\"value\" in descriptor) descriptor.writable = true;\n      Object.defineProperty(target, descriptor.key, descriptor);\n    }\n  }\n  return function (Constructor, protoProps, staticProps) {\n    if (protoProps) defineProperties(Constructor.prototype, protoProps);\n    if (staticProps) defineProperties(Constructor, staticProps);\n    return Constructor;\n  };\n})();\n\nvar _sat = __webpack_require__(/*! sat */ \"./node_modules/sat/SAT.js\");\n\nvar _sat2 = _interopRequireDefault(_sat);\n\nvar _model = __webpack_require__(/*! ../model */ \"./dist/model.js\");\n\nvar _utils = __webpack_require__(/*! ../utils */ \"./dist/utils.js\");\n\nfunction _interopRequireDefault(obj) {\n  return obj && obj.__esModule ? obj : { default: obj };\n}\n\nfunction _classCallCheck(instance, Constructor) {\n  if (!(instance instanceof Constructor)) {\n    throw new TypeError(\"Cannot call a class as a function\");\n  }\n}\n\nfunction _possibleConstructorReturn(self, call) {\n  if (!self) {\n    throw new ReferenceError(\n      \"this hasn't been initialised - super() hasn't been called\"\n    );\n  }\n  return call && (typeof call === \"object\" || typeof call === \"function\")\n    ? call\n    : self;\n}\n\nfunction _inherits(subClass, superClass) {\n  if (typeof superClass !== \"function\" && superClass !== null) {\n    throw new TypeError(\n      \"Super expression must either be null or a function, not \" +\n        typeof superClass\n    );\n  }\n  subClass.prototype = Object.create(superClass && superClass.prototype, {\n    constructor: {\n      value: subClass,\n      enumerable: false,\n      writable: true,\n      configurable: true,\n    },\n  });\n  if (superClass)\n    Object.setPrototypeOf\n      ? Object.setPrototypeOf(subClass, superClass)\n      : (subClass.__proto__ = superClass);\n}\n\n/**\n * collider - polygon\n */\nvar Polygon = (exports.Polygon = (function (_SAT$Polygon) {\n  _inherits(Polygon, _SAT$Polygon);\n\n  /**\n   * collider - polygon\n   * @param {Vector} position {x, y}\n   * @param {Vector[]} points\n   */\n  function Polygon(position, points) {\n    _classCallCheck(this, Polygon);\n\n    var _this = _possibleConstructorReturn(\n      this,\n      (Polygon.__proto__ || Object.getPrototypeOf(Polygon)).call(\n        this,\n        (0, _utils.ensureVectorPoint)(position),\n        (0, _utils.ensurePolygonPoints)(points)\n      )\n    );\n\n    _this.type = _model.Types.Polygon;\n    _this.updateAABB();\n    return _this;\n  }\n  /**\n   * Updates Bounding Box of collider\n   */\n\n  _createClass(Polygon, [\n    {\n      key: \"updateAABB\",\n      value: function updateAABB() {\n        var _getAABBAsBox = this.getAABBAsBox(),\n          pos = _getAABBAsBox.pos,\n          w = _getAABBAsBox.w,\n          h = _getAABBAsBox.h;\n\n        this.minX = pos.x;\n        this.minY = pos.y;\n        this.maxX = pos.x + w;\n        this.maxY = pos.y + h;\n      },\n      /**\n       * Draws collider on a CanvasRenderingContext2D's current path\n       * @param {CanvasRenderingContext2D} context The canvas context to draw on\n       */\n    },\n    {\n      key: \"draw\",\n      value: function draw(context) {\n        var _this2 = this;\n\n        this.calcPoints.forEach(function (point, index) {\n          var posX = _this2.pos.x + point.x;\n          var posY = _this2.pos.y + point.y;\n          if (!index) {\n            if (_this2.calcPoints.length === 1) {\n              context.arc(posX, posY, 1, 0, Math.PI * 2);\n            } else {\n              context.moveTo(posX, posY);\n            }\n          } else {\n            context.lineTo(posX, posY);\n          }\n        });\n        if (this.calcPoints.length > 2) {\n          context.lineTo(\n            this.pos.x + this.calcPoints[0].x,\n            this.pos.y + this.calcPoints[0].y\n          );\n        }\n      },\n    },\n  ]);\n\n  return Polygon;\n})(_sat2.default.Polygon));\n\n\n//# sourceURL=webpack://detect-collisions/./dist/bodies/polygon.js?");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}));\nexports.Polygon = undefined;\n\nvar _createClass = (function () {\n  function defineProperties(target, props) {\n    for (var i = 0; i < props.length; i++) {\n      var descriptor = props[i];\n      descriptor.enumerable = descriptor.enumerable || false;\n      descriptor.configurable = true;\n      if (\"value\" in descriptor) descriptor.writable = true;\n      Object.defineProperty(target, descriptor.key, descriptor);\n    }\n  }\n  return function (Constructor, protoProps, staticProps) {\n    if (protoProps) defineProperties(Constructor.prototype, protoProps);\n    if (staticProps) defineProperties(Constructor, staticProps);\n    return Constructor;\n  };\n})();\n\nvar _sat = __webpack_require__(/*! sat */ \"./node_modules/sat/SAT.js\");\n\nvar _sat2 = _interopRequireDefault(_sat);\n\nvar _model = __webpack_require__(/*! ../model */ \"./dist/model.js\");\n\nvar _utils = __webpack_require__(/*! ../utils */ \"./dist/utils.js\");\n\nfunction _interopRequireDefault(obj) {\n  return obj && obj.__esModule ? obj : { default: obj };\n}\n\nfunction _toConsumableArray(arr) {\n  if (Array.isArray(arr)) {\n    for (var i = 0, arr2 = Array(arr.length); i < arr.length; i++) {\n      arr2[i] = arr[i];\n    }\n    return arr2;\n  } else {\n    return Array.from(arr);\n  }\n}\n\nfunction _classCallCheck(instance, Constructor) {\n  if (!(instance instanceof Constructor)) {\n    throw new TypeError(\"Cannot call a class as a function\");\n  }\n}\n\nfunction _possibleConstructorReturn(self, call) {\n  if (!self) {\n    throw new ReferenceError(\n      \"this hasn't been initialised - super() hasn't been called\"\n    );\n  }\n  return call && (typeof call === \"object\" || typeof call === \"function\")\n    ? call\n    : self;\n}\n\nfunction _inherits(subClass, superClass) {\n  if (typeof superClass !== \"function\" && superClass !== null) {\n    throw new TypeError(\n      \"Super expression must either be null or a function, not \" +\n        typeof superClass\n    );\n  }\n  subClass.prototype = Object.create(superClass && superClass.prototype, {\n    constructor: {\n      value: subClass,\n      enumerable: false,\n      writable: true,\n      configurable: true,\n    },\n  });\n  if (superClass)\n    Object.setPrototypeOf\n      ? Object.setPrototypeOf(subClass, superClass)\n      : (subClass.__proto__ = superClass);\n}\n\n/**\n * collider - polygon\n */\nvar Polygon = (exports.Polygon = (function (_SAT$Polygon) {\n  _inherits(Polygon, _SAT$Polygon);\n\n  /**\n   * collider - polygon\n   * @param {Vector} position {x, y}\n   * @param {Vector[]} points\n   */\n  function Polygon(position, points) {\n    _classCallCheck(this, Polygon);\n\n    var _this = _possibleConstructorReturn(\n      this,\n      (Polygon.__proto__ || Object.getPrototypeOf(Polygon)).call(\n        this,\n        (0, _utils.ensureVectorPoint)(position),\n        (0, _utils.ensurePolygonPoints)(points)\n      )\n    );\n\n    _this.type = _model.Types.Polygon;\n    _this.updateAABB();\n    return _this;\n  }\n  /**\n   * Updates Bounding Box of collider\n   */\n\n  _createClass(Polygon, [\n    {\n      key: \"updateAABB\",\n      value: function updateAABB() {\n        var _getAABBAsBox = this.getAABBAsBox(),\n          pos = _getAABBAsBox.pos,\n          w = _getAABBAsBox.w,\n          h = _getAABBAsBox.h;\n\n        this.minX = pos.x;\n        this.minY = pos.y;\n        this.maxX = pos.x + w;\n        this.maxY = pos.y + h;\n      },\n      /**\n       * Draws collider on a CanvasRenderingContext2D's current path\n       * @param {CanvasRenderingContext2D} context The canvas context to draw on\n       */\n    },\n    {\n      key: \"draw\",\n      value: function draw(context) {\n        var _this2 = this;\n\n        []\n          .concat(_toConsumableArray(this.calcPoints), [this.calcPoints[0]])\n          .forEach(function (point, index) {\n            var toX = _this2.pos.x + point.x;\n            var toY = _this2.pos.y + point.y;\n            var prev =\n              _this2.calcPoints[index - 1] ||\n              _this2.calcPoints[_this2.calcPoints.length - 1];\n            if (!index) {\n              if (_this2.calcPoints.length === 1) {\n                context.arc(toX, toY, 1, 0, Math.PI * 2);\n              } else {\n                context.moveTo(toX, toY);\n              }\n            } else if (_this2.calcPoints.length > 1) {\n              if (_this2.isTrigger) {\n                var fromX = _this2.pos.x + prev.x;\n                var fromY = _this2.pos.y + prev.y;\n                (0, _utils.dashLineTo)(context, fromX, fromY, toX, toY);\n              } else {\n                context.lineTo(toX, toY);\n              }\n            }\n          });\n      },\n    },\n  ]);\n\n  return Polygon;\n})(_sat2.default.Polygon));\n\n\n//# sourceURL=webpack://detect-collisions/./dist/bodies/polygon.js?");
 
 /***/ }),
 
@@ -93,7 +93,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}))
 /***/ ((__unused_webpack_module, exports, __webpack_require__) => {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}));\nexports.createBox = createBox;\nexports.ensureVectorPoint = ensureVectorPoint;\nexports.ensurePolygonPoints = ensurePolygonPoints;\nexports.clockwise = clockwise;\n\nvar _sat = __webpack_require__(/*! sat */ \"./node_modules/sat/SAT.js\");\n\nvar _sat2 = _interopRequireDefault(_sat);\n\nfunction _interopRequireDefault(obj) {\n  return obj && obj.__esModule ? obj : { default: obj };\n}\n\n/**\n * creates box polygon points\n * @param {number} width\n * @param {number} height\n * @returns SAT.Vector\n */\nfunction createBox(width, height) {\n  return [\n    new _sat2.default.Vector(),\n    new _sat2.default.Vector(width, 0),\n    new _sat2.default.Vector(width, height),\n    new _sat2.default.Vector(0, height),\n  ];\n}\n/**\n * ensure returns a SAT.Vector\n * @param {SAT.Vector} point\n */\nfunction ensureVectorPoint(point) {\n  return point instanceof _sat2.default.Vector\n    ? point\n    : new _sat2.default.Vector(point.x, point.y);\n}\n/**\n * ensure correct counterclockwise points\n * @param {SAT.Vector[]} points\n */\nfunction ensurePolygonPoints(points) {\n  return (clockwise(points) ? points.reverse() : points).map(ensureVectorPoint);\n}\n/**\n * check direction of polygon\n * @param {SAT.Vector[]} points\n */\nfunction clockwise(points) {\n  var sum = 0;\n  for (var i = 0; i < points.length; i++) {\n    var v1 = points[i];\n    var v2 = points[(i + 1) % points.length];\n    sum += (v2.x - v1.x) * (v2.y + v1.y);\n  }\n  return sum > 0;\n}\n\n\n//# sourceURL=webpack://detect-collisions/./dist/utils.js?");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", ({\n  value: true,\n}));\nexports.createBox = createBox;\nexports.ensureVectorPoint = ensureVectorPoint;\nexports.ensurePolygonPoints = ensurePolygonPoints;\nexports.clockwise = clockwise;\nexports.dashLineTo = dashLineTo;\n\nvar _sat = __webpack_require__(/*! sat */ \"./node_modules/sat/SAT.js\");\n\nvar _sat2 = _interopRequireDefault(_sat);\n\nfunction _interopRequireDefault(obj) {\n  return obj && obj.__esModule ? obj : { default: obj };\n}\n\n/**\n * creates box polygon points\n * @param {number} width\n * @param {number} height\n * @returns SAT.Vector\n */\nfunction createBox(width, height) {\n  return [\n    new _sat2.default.Vector(),\n    new _sat2.default.Vector(width, 0),\n    new _sat2.default.Vector(width, height),\n    new _sat2.default.Vector(0, height),\n  ];\n}\n/**\n * ensure returns a SAT.Vector\n * @param {SAT.Vector} point\n */\nfunction ensureVectorPoint(point) {\n  return point instanceof _sat2.default.Vector\n    ? point\n    : new _sat2.default.Vector(point.x, point.y);\n}\n/**\n * ensure correct counterclockwise points\n * @param {SAT.Vector[]} points\n */\nfunction ensurePolygonPoints(points) {\n  return (clockwise(points) ? points.reverse() : points).map(ensureVectorPoint);\n}\n/**\n * check direction of polygon\n * @param {SAT.Vector[]} points\n */\nfunction clockwise(points) {\n  var sum = 0;\n  for (var i = 0; i < points.length; i++) {\n    var v1 = points[i];\n    var v2 = points[(i + 1) % points.length];\n    sum += (v2.x - v1.x) * (v2.y + v1.y);\n  }\n  return sum > 0;\n}\n/**\n * draws dashed line on canvas context\n * @param {CanvasRenderingContext2D} context\n * @param {number} fromX\n * @param {number} fromY\n * @param {number} toX\n * @param {number} toY\n * @param {number?} dash\n * @param {number?} gap\n */\nfunction dashLineTo(context, fromX, fromY, toX, toY) {\n  var dash =\n    arguments.length > 5 && arguments[5] !== undefined ? arguments[5] : 2;\n  var gap =\n    arguments.length > 6 && arguments[6] !== undefined ? arguments[6] : 4;\n\n  var xDiff = toX - fromX;\n  var yDiff = toY - fromY;\n  var arc = Math.atan2(yDiff, xDiff);\n  var offsetX = Math.cos(arc);\n  var offsetY = Math.sin(arc);\n  var posX = fromX;\n  var posY = fromY;\n  var distance = Math.hypot(xDiff, yDiff);\n  while (distance > 0) {\n    var step = Math.min(distance, dash);\n    context.moveTo(posX, posY);\n    context.lineTo(posX + offsetX * step, posY + offsetY * step);\n    posX += offsetX * (dash + gap);\n    posY += offsetY * (dash + gap);\n    distance -= dash + gap;\n  }\n}\n\n\n//# sourceURL=webpack://detect-collisions/./dist/utils.js?");
 
 /***/ }),
 
