@@ -1,6 +1,6 @@
 import SAT from "sat";
 import { Types } from "../model";
-import { ensureVectorPoint, ensurePolygonPoints } from "../utils";
+import { ensureVectorPoint, ensurePolygonPoints, dashLineTo } from "../utils";
 /**
  * collider - polygon
  */
@@ -30,23 +30,29 @@ export class Polygon extends SAT.Polygon {
      * @param {CanvasRenderingContext2D} context The canvas context to draw on
      */
     draw(context) {
-        this.calcPoints.forEach((point, index) => {
-            const posX = this.pos.x + point.x;
-            const posY = this.pos.y + point.y;
+        [...this.calcPoints, this.calcPoints[0]].forEach((point, index) => {
+            const toX = this.pos.x + point.x;
+            const toY = this.pos.y + point.y;
+            const prev = this.calcPoints[index - 1] ||
+                this.calcPoints[this.calcPoints.length - 1];
             if (!index) {
                 if (this.calcPoints.length === 1) {
-                    context.arc(posX, posY, 1, 0, Math.PI * 2);
+                    context.arc(toX, toY, 1, 0, Math.PI * 2);
                 }
                 else {
-                    context.moveTo(posX, posY);
+                    context.moveTo(toX, toY);
                 }
             }
-            else {
-                context.lineTo(posX, posY);
+            else if (this.calcPoints.length > 1) {
+                if (this.isTrigger) {
+                    const fromX = this.pos.x + prev.x;
+                    const fromY = this.pos.y + prev.y;
+                    dashLineTo(context, fromX, fromY, toX, toY);
+                }
+                else {
+                    context.lineTo(toX, toY);
+                }
             }
         });
-        if (this.calcPoints.length > 2) {
-            context.lineTo(this.pos.x + this.calcPoints[0].x, this.pos.y + this.calcPoints[0].y);
-        }
     }
 }
